+++ src/werkzeug/datastructures/etag.py
@@ -8,6 +8,92 @@ if t.TYPE_CHECKING:
     import typing_extensions as te
 
 
+class ETag:
+    """A parsed ETag value.
+
+    Set :attr:`.Response.etag` to an instance to set the header. Modifying the
+    instance will update the header.
+
+    .. versionadded:: 3.2
+    """
+
+    _on_update: t.Callable[[ETag], None] | None = None
+
+    def __init__(self, value: str, weak: bool = False) -> None:
+        self._value = value
+        self._weak = weak
+
+    def _trigger_on_update(self) -> None:
+        if self._on_update is not None:
+            self._on_update(self)
+
+    @property
+    def value(self) -> str:
+        """The unquoted value."""
+        return self._value
+
+    @value.setter
+    def value(self, value: str) -> None:
+        self._value = value
+        self._trigger_on_update()
+
+    @property
+    def weak(self) -> bool:
+        """Whether the weak marker is present."""
+        return self._weak
+
+    @weak.setter
+    def weak(self, value: bool) -> None:
+        self._weak = value
+        self._trigger_on_update()
+
+    @classmethod
+    def from_header(cls, value: str | None) -> te.Self | None:
+        """Parse a quoted ETag value and create an instance of this class, or
+        ``None`` if the value is empty or invalid.
+        """
+
+        if not value:
+            return None
+
+        weak = False
+        start = 0
+
+        if value.startswith(("W/", "w/")):
+            weak = True
+            start = 2
+
+        if not (value.startswith('"', start) and value.endswith('"', start)):
+            # invalid, value must be quoted
+            return None
+
+        return cls(value[start + 1 : -1], weak)
+
+    def to_header(self) -> str:
+        if not self.value or '"' in self.value:
+            return ""
+
+        if self.weak:
+            return f'W/"{self.value}"'
+
+        return f'"{self.value}"'
+
+    def __bool__(self) -> bool:
+        return bool(self._value)
+
+    def __eq__(self, value: object, /) -> bool:
+        if not isinstance(value, type(self)):
+            return NotImplemented
+
+        return self._value == value._value and self._weak == value._weak
+
+    def __str__(self) -> str:
+        return self.to_header()
+
+    def __repr__(self) -> str:
+        return f"<ETag {self.value!r} {'weak' if self.weak else 'strong'}>"
+
+
 _etag_re = re.compile(
     r"""
     [ \t]*  # ignore leading space
@@ -20,7 +106,7 @@ _etag_re = re.compile(
 )
 
 
-class ETags(cabc.Collection[str]):
+class ETagSet(cabc.Collection[str]):
     """A parsed ``If-Match`` or ``If-None-Match`` header.
 
     :attr:`.Request.if_match` and :attr:`.Request.if_none_match` return an
@@ -29,6 +115,9 @@ class ETags(cabc.Collection[str]):
     :param strong_etags: Unquoted values that were not marked weak.
     :param weak_etags: Unquoted values that were marked weak.
     :param star_tag: Whether ``*`` is present in the header value.
+
+    .. versionchanged:: 3.2
+        Renamed from ``ETags``.
     """
 
     def __init__(
@@ -102,17 +191,13 @@ class ETags(cabc.Collection[str]):
 
         :param etag: The raw, quoted value to check.
         """
-        from ..http import unquote_etag
-
-        value, weak = unquote_etag(etag)
-
-        if value is None or weak is None:
+        if (value := ETag.from_header(etag)) is None:
             return False
 
-        if weak:
-            return self.contains_weak(value)
+        if value.weak:
+            return self.contains_weak(value.value)
 
-        return self.contains(value)
+        return self.contains(value.value)
 
     @classmethod
     def from_header(cls, value: str | None) -> te.Self:
@@ -205,3 +290,20 @@ class ETags(cabc.Collection[str]):
 
     def __repr__(self) -> str:
         return f"<{type(self).__name__} {str(self)!r}>"
+
+
+if not t.TYPE_CHECKING:
+
+    def __getattr__(name: str) -> t.Any:
+        if name == "ETags":
+            import warnings
+
+            warnings.warn(
+                "'ETags' has been renamed to 'ETagSet'. The old name is deprecated and"
+                " will be removed in Werkzeug 3.3.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+            return ETagSet
+
+        raise AttributeError(name)